Consolidate python functions, add rosters/stats
