+++ csgo.bat
@@ -1,14 +1,7 @@
 echo off
 
-C:\Anaconda2\python.exe -W ignore %~dp0\python\scrape\hltv_scrape_teamrank.py
-C:\Anaconda2\python.exe -W ignore %~dp0\python\scrape\hltv_scrape_event.py
-C:\Anaconda2\python.exe -W ignore %~dp0\python\scrape\hltv_scrape_match.py
-C:\Anaconda2\python.exe -W ignore %~dp0\python\scrape\hltv_scrape_demo.py
-
-::C:\Anaconda2\python.exe -W ignore %~dp0\python\parse\rar_to_demo.py
-::C:\Anaconda2\python.exe -W ignore %~dp0\python\parse\demo_to_json.py
-::C:\Anaconda2\python.exe -W ignore %~dp0\python\parse\json_to_csv.py
-
+C:\Anaconda2\python.exe -W ignore %~dp0\python\hltv_scrape.py
+::C:\Anaconda2\python.exe -W ignore %~dp0\python\demo_parse.py
 ::C:\Anaconda2\python.exe -W ignore %~dp0\python\other\team_name_match.py
 
 for /f %%a in ('psql -U postgres -c "select 1 as result from pg_database where datname='esports'" -t') do set /a check=%%a
@@ -22,6 +15,8 @@ psql -U postgres -d esports -qf sql/hltv_events.sql
 psql -U postgres -d esports -qf sql/hltv_event_team_places.sql
 psql -U postgres -d esports -qf sql/hltv_match_info.sql
 psql -U postgres -d esports -qf sql/hltv_map_results.sql
+psql -U postgres -d esports -qf sql/hltv_match_stats.sql
+psql -U postgres -d esports -qf sql/hltv_team_rosters.sql
 psql -U postgres -d esports -qf sql/hltv_vetos.sql
 ::psql -U postgres -d esports -qf sql/demo_info.sql
 ::psql -U postgres -d esports -qf sql/demo_players.sql
