+++ csv/hltv_event_team_places.csv
@@ -1,0 +1,98 @@
+/events/2570/dreamhack-open-valencia-2017,07/15/2017,/team/4411/nip,1,50000
+/events/2570/dreamhack-open-valencia-2017,07/15/2017,/team/7613/red-reserve,2,20000
+/events/2570/dreamhack-open-valencia-2017,07/15/2017,/team/5991/envyus,3,10000
+/events/2570/dreamhack-open-valencia-2017,07/15/2017,/team/7175/heroic,3,10000
+/events/2570/dreamhack-open-valencia-2017,07/15/2017,/team/5974/clg,5,3000
+/events/2570/dreamhack-open-valencia-2017,07/15/2017,/team/7339/kpi,5,3000
+/events/2570/dreamhack-open-valencia-2017,07/15/2017,/team/6673/nrg,7,2000
+/events/2570/dreamhack-open-valencia-2017,07/15/2017,/team/7557/misfits,7,2000
+/events/2635/esl-one-cologne-2017,07/09/2017,/team/6137/sk,1,100000
+/events/2635/esl-one-cologne-2017,07/09/2017,/team/5752/cloud9,2,40000
+/events/2635/esl-one-cologne-2017,07/09/2017,/team/4608/natus-vincere,3,20000
+/events/2635/esl-one-cologne-2017,07/09/2017,/team/6667/faze,3,20000
+/events/2635/esl-one-cologne-2017,07/09/2017,/team/4411/nip,5,10000
+/events/2635/esl-one-cologne-2017,07/09/2017,/team/5973/liquid,5,10000
+/events/2635/esl-one-cologne-2017,07/09/2017,/team/5995/g2,5,10000
+/events/2635/esl-one-cologne-2017,07/09/2017,/team/6615/optic,5,10000
+/events/2635/esl-one-cologne-2017,07/09/2017,/team/4494/mousesports,9,6000
+/events/2635/esl-one-cologne-2017,07/09/2017,/team/4991/fnatic,9,6000
+/events/2635/esl-one-cologne-2017,07/09/2017,/team/7533/north,9,6000
+/events/2635/esl-one-cologne-2017,07/09/2017,/team/7010/immortals,12,3000
+/events/2635/esl-one-cologne-2017,07/09/2017,/team/7175/heroic,12,3000
+/events/2635/esl-one-cologne-2017,07/09/2017,/team/5929/space-soldiers,12,3000
+/events/2635/esl-one-cologne-2017,07/09/2017,/team/4863/tyloo,15,1500
+/events/2635/esl-one-cologne-2017,07/09/2017,/team/5378/virtuspro,15,1500
+/events/2721/pgl-major-krakow-2017-main-qualifier,07/02/2017,/team/4494/mousesports,1,0
+/events/2721/pgl-major-krakow-2017-main-qualifier,07/02/2017,/team/5995/g2,1,0
+/events/2721/pgl-major-krakow-2017-main-qualifier,07/02/2017,/team/5395/penta,3,0
+/events/2721/pgl-major-krakow-2017-main-qualifier,07/02/2017,/team/5752/cloud9,3,0
+/events/2721/pgl-major-krakow-2017-main-qualifier,07/02/2017,/team/7532/big,3,0
+/events/2721/pgl-major-krakow-2017-main-qualifier,07/02/2017,/team/6094/vega-squadron,6,0
+/events/2721/pgl-major-krakow-2017-main-qualifier,07/02/2017,/team/7010/immortals,6,0
+/events/2721/pgl-major-krakow-2017-main-qualifier,07/02/2017,/team/5988/flipsid3,6,0
+/events/2721/pgl-major-krakow-2017-main-qualifier,07/02/2017,/team/5310/hellraisers,9,0
+/events/2721/pgl-major-krakow-2017-main-qualifier,07/02/2017,/team/5422/dignitas,9,0
+/events/2721/pgl-major-krakow-2017-main-qualifier,07/02/2017,/team/5973/liquid,9,0
+/events/2721/pgl-major-krakow-2017-main-qualifier,07/02/2017,/team/4863/tyloo,12,0
+/events/2721/pgl-major-krakow-2017-main-qualifier,07/02/2017,/team/6211/renegades,12,0
+/events/2721/pgl-major-krakow-2017-main-qualifier,07/02/2017,/team/6902/godsent,12,0
+/events/2721/pgl-major-krakow-2017-main-qualifier,07/02/2017,/team/6615/optic,15,0
+/events/2721/pgl-major-krakow-2017-main-qualifier,07/02/2017,/team/7145/tengri,15,0
+/events/2763/ecs-season-3-finals,06/25/2017,/team/6137/sk,1,250000
+/events/2763/ecs-season-3-finals,06/25/2017,/team/6667/faze,2,120000
+/events/2763/ecs-season-3-finals,06/25/2017,/team/5752/cloud9,3,65000
+/events/2763/ecs-season-3-finals,06/25/2017,/team/6665/astralis,3,65000
+/events/2763/ecs-season-3-finals,06/25/2017,/team/5973/liquid,5,45000
+/events/2763/ecs-season-3-finals,06/25/2017,/team/5995/g2,5,45000
+/events/2763/ecs-season-3-finals,06/25/2017,/team/4991/fnatic,7,35000
+/events/2763/ecs-season-3-finals,06/25/2017,/team/6615/optic,7,35000
+/events/2569/dreamhack-open-summer-2017,06/19/2017,/team/6137/sk,1,50000
+/events/2569/dreamhack-open-summer-2017,06/19/2017,/team/4991/fnatic,2,20000
+/events/2569/dreamhack-open-summer-2017,06/19/2017,/team/5974/clg,3,10000
+/events/2569/dreamhack-open-summer-2017,06/19/2017,/team/7010/immortals,3,10000
+/events/2569/dreamhack-open-summer-2017,06/19/2017,/team/4494/mousesports,5,3000
+/events/2569/dreamhack-open-summer-2017,06/19/2017,/team/6651/gambit,5,3000
+/events/2569/dreamhack-open-summer-2017,06/19/2017,/team/5752/cloud9,7,2000
+/events/2569/dreamhack-open-summer-2017,06/19/2017,/team/6978/singularity,7,2000
+/events/2889/adrenaline-cyber-league-2017,06/18/2017,/team/5378/virtuspro,1,65000
+/events/2889/adrenaline-cyber-league-2017,06/18/2017,/team/4608/natus-vincere,2,25000
+/events/2889/adrenaline-cyber-league-2017,06/18/2017,/team/5310/hellraisers,3,5000
+/events/2889/adrenaline-cyber-league-2017,06/18/2017,/team/5988/flipsid3,3,5000
+/events/2723/europe-minor-pgl-major-krakow-2017,06/18/2017,/team/7532/big,1,30000
+/events/2723/europe-minor-pgl-major-krakow-2017,06/18/2017,/team/5395/penta,2,15000
+/events/2723/europe-minor-pgl-major-krakow-2017,06/18/2017,/team/5422/dignitas,3,5000
+/events/2723/europe-minor-pgl-major-krakow-2017,06/18/2017,/team/5991/envyus,4,0
+/events/2723/europe-minor-pgl-major-krakow-2017,06/18/2017,/team/4674/ldlc,5,0
+/events/2723/europe-minor-pgl-major-krakow-2017,06/18/2017,/team/7990/ballistix,5,0
+/events/2723/europe-minor-pgl-major-krakow-2017,06/18/2017,/team/6134/kinguin,7,0
+/events/2723/europe-minor-pgl-major-krakow-2017,06/18/2017,/team/7331/igamecom,7,0
+/events/2853/eleague-clash-for-cash,06/16/2017,/team/6665/astralis,1,250000
+/events/2853/eleague-clash-for-cash,06/16/2017,/team/5378/virtuspro,2,0
+/events/2722/americas-minor-pgl-major-krakow-2017,06/11/2017,/team/5752/cloud9,1,30000
+/events/2722/americas-minor-pgl-major-krakow-2017,06/11/2017,/team/7010/immortals,2,15000
+/events/2722/americas-minor-pgl-major-krakow-2017,06/11/2017,/team/5974/clg,3,5000
+/events/2722/americas-minor-pgl-major-krakow-2017,06/11/2017,/team/6290/luminosity,4,0
+/events/2722/americas-minor-pgl-major-krakow-2017,06/11/2017,/team/4773/pain,5,0
+/events/2722/americas-minor-pgl-major-krakow-2017,06/11/2017,/team/6673/nrg,5,0
+/events/2722/americas-minor-pgl-major-krakow-2017,06/11/2017,/team/5005/complexity,7,0
+/events/2722/americas-minor-pgl-major-krakow-2017,06/11/2017,/team/7557/misfits,7,0
+/events/2725/asia-minor-pgl-major-krakow-2017,06/04/2017,/team/6211/renegades,1,30000
+/events/2725/asia-minor-pgl-major-krakow-2017,06/04/2017,/team/4863/tyloo,2,15000
+/events/2725/asia-minor-pgl-major-krakow-2017,06/04/2017,/team/4791/immunity,3,5000
+/events/2725/asia-minor-pgl-major-krakow-2017,06/04/2017,/team/7660/flash,4,0
+/events/2725/asia-minor-pgl-major-krakow-2017,06/04/2017,/team/6248/themongolz,5,0
+/events/2725/asia-minor-pgl-major-krakow-2017,06/04/2017,/team/6774/signature,5,0
+/events/2725/asia-minor-pgl-major-krakow-2017,06/04/2017,/team/7576/7642,7,0
+/events/2725/asia-minor-pgl-major-krakow-2017,06/04/2017,/team/7728/spotnet,7,0
+/events/2557/esl-pro-league-season-5-finals,06/04/2017,/team/5995/g2,1,225000
+/events/2557/esl-pro-league-season-5-finals,06/04/2017,/team/7533/north,2,100000
+/events/2557/esl-pro-league-season-5-finals,06/04/2017,/team/5973/liquid,3,60000
+/events/2557/esl-pro-league-season-5-finals,06/04/2017,/team/6137/sk,3,60000
+/events/2557/esl-pro-league-season-5-finals,06/04/2017,/team/4494/mousesports,5,45000
+/events/2557/esl-pro-league-season-5-finals,06/04/2017,/team/5991/envyus,5,45000
+/events/2557/esl-pro-league-season-5-finals,06/04/2017,/team/6615/optic,7,40000
+/events/2557/esl-pro-league-season-5-finals,06/04/2017,/team/4991/fnatic,7,40000
+/events/2557/esl-pro-league-season-5-finals,06/04/2017,/team/4608/natus-vincere,9,35000
+/events/2557/esl-pro-league-season-5-finals,06/04/2017,/team/7010/immortals,9,35000
+/events/2557/esl-pro-league-season-5-finals,06/04/2017,/team/6673/nrg,11,32500
+/events/2557/esl-pro-league-season-5-finals,06/04/2017,/team/5752/cloud9,11,32500
